+++ scripts/governance/addSeries/addEthSeries/addEthSeries.sh
@@ -6,11 +6,11 @@ export CONF=$PWD/$HERE/addEthSeries.mainnet.config
 RUN="npx hardhat run --network localhost"
 
 # Add funds to the timelock
-# $RUN $HERE/loadTimelock.ts
-# 
-# $RUN $HERE/../../newEnvironment/deployFYTokens.ts # deploy fyTokens
-# $RUN $HERE/../../newEnvironment/deployPools.ts # deploy pools
-# $RUN $HERE/../../newEnvironment/deployStrategies.ts # deploy strategies
+$RUN $HERE/loadTimelock.ts
+
+$RUN $HERE/../../newEnvironment/deployFYTokens.ts # deploy fyTokens
+$RUN $HERE/../../newEnvironment/deployPools.ts # deploy pools
+$RUN $HERE/../../newEnvironment/deployStrategies.ts # deploy strategies
 
 $RUN $HERE/addEthSeries.ts
 $RUN $HERE/addEthSeries.ts
